requirements.txt
alembic==1.14.1
backports.entry-points-selectable==1.3.0
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
honcho==2.0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==2.1.5
packaging==24.2
psycopg2==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.40
SQLAlchemy-serializer==1.4.12
typing_extensions==4.13.0
Werkzeug==3.0.6
zipp==3.20.2
